Updates for AES armv8.
