+++ src/common/aes/aes.h
@@ -133,21 +133,6 @@ void OQS_AES256_ECB_enc_sch(const uint8_t *plaintext, const size_t plaintext_len
 void OQS_AES256_CTR_inc_stream_iv(const uint8_t *iv, size_t iv_len, const void *ctx, uint8_t *out, size_t out_len);
 
 /**
- * AES counter mode keystream generator.  A context generated by
- * OQS_AES256_CTR_inc_init() is passed rather then a key.
- *
- * Handles an 8-byte IV passed as a 64-bit unsigned integer,
- * counter bytes are initialized to zero.
- *
- * @param iv       64-bit initialization vector.
- * @param ctx      Abstract data structure for a key schedule.
- * @param out      Pointer to a block of memory which is big enough to contain out_len bytes; the result will be written here.
- * @param out_blks Number of output blocks (16 bytes) to generate.
- */
-void OQS_AES256_CTR_inc_stream_ivu64_blks(uint64_t iv, void *ctx, uint8_t *out, size_t out_blks);
-
-
-/**
  * AES counter mode keystream generator. A context generated by
  * OQS_AES256_CTR_inc_init() and OQS_AES256_CTR_inc_iv() is passed
  * rather than a key and an IV. The counter is internally updated, which allows
