+++ src/common/aes/aes_c.c
@@ -53,6 +53,10 @@ typedef struct {
 	uint8_t iv[AES_BLOCKBYTES];
 } aes256ctx;
 
+typedef struct {
+	uint32_t sk_exp[60];
+	uint8_t iv[16];
+} aes256ctx_nobitslice;
 
 static inline uint32_t br_dec32le(const unsigned char *src) {
 	return (uint32_t)src[0]
@@ -416,31 +420,6 @@ static void br_aes_ct64_keysched(uint64_t *comp_skey, const unsigned char *key, 
 	}
 }
 
-static void aes_keysched_no_bitslice(uint32_t *skey, const unsigned char *key, unsigned int key_len) {
-	unsigned int i, j, k, nk, nkf;
-	uint32_t tmp;
-	unsigned nrounds = 10 + ((key_len - 16) >> 2);
-
-	nk = (key_len >> 2);
-	nkf = ((nrounds + 1) << 2);
-	br_range_dec32le(skey, (key_len >> 2), key);
-	tmp = skey[(key_len >> 2) - 1];
-	for (i = nk, j = 0, k = 0; i < nkf; i ++) {
-		if (j == 0) {
-			tmp = (tmp << 24) | (tmp >> 8);
-			tmp = sub_word(tmp) ^ Rcon[k];
-		} else if (nk > 6 && j == 4) {
-			tmp = sub_word(tmp);
-		}
-		tmp ^= skey[i - nk];
-		skey[i] = tmp;
-		if (++ j == nk) {
-			j = 0;
-			k ++;
-		}
-	}
-}
-
 static void br_aes_ct64_skey_expand(uint64_t *skey, const uint64_t *comp_skey, unsigned int nrounds) {
 	unsigned u, v, n;
 
@@ -681,6 +660,38 @@ void oqs_aes256_load_schedule_c(const uint8_t *key, void **_schedule) {
 	br_aes_ct64_skey_expand(ctx->sk_exp, skey, 14);
 }
 
+static void aes_keysched_no_bitslice(uint32_t *skey, const unsigned char *key, unsigned int key_len) {
+	unsigned int i, j, k, nk, nkf;
+	uint32_t tmp;
+	unsigned nrounds = 10 + ((key_len - 16) >> 2);
+
+	nk = (key_len >> 2);
+	nkf = ((nrounds + 1) << 2);
+	br_range_dec32le(skey, (key_len >> 2), key);
+	tmp = skey[(key_len >> 2) - 1];
+	for (i = nk, j = 0, k = 0; i < nkf; i ++) {
+		if (j == 0) {
+			tmp = (tmp << 24) | (tmp >> 8);
+			tmp = sub_word(tmp) ^ Rcon[k];
+		} else if (nk > 6 && j == 4) {
+			tmp = sub_word(tmp);
+		}
+		tmp ^= skey[i - nk];
+		skey[i] = tmp;
+		if (++ j == nk) {
+			j = 0;
+			k ++;
+		}
+	}
+}
+
+void oqs_aes256_load_schedule_no_bitslice(const uint8_t *key, void **_schedule) {
+	*_schedule = malloc(sizeof(aes256ctx_nobitslice));
+	assert(*_schedule != NULL);
+	uint32_t *schedule = ((aes256ctx_nobitslice *) *_schedule)->sk_exp;
+	aes_keysched_no_bitslice(schedule, (const unsigned char *) key, 32);
+}
+
 void oqs_aes256_load_iv_c(const uint8_t *iv, size_t iv_len, void *_schedule) {
 	aes256ctx *ctx = _schedule;
 	if (iv_len == 12) {
@@ -714,13 +725,6 @@ void oqs_aes128_load_schedule_no_bitslice(const uint8_t *key, void **_schedule) 
 	aes_keysched_no_bitslice(schedule, (const unsigned char *) key, 16);
 }
 
-void oqs_aes256_load_schedule_no_bitslice(const uint8_t *key, void **_schedule) {
-	*_schedule = malloc(60 * sizeof(int));
-	assert(*_schedule != NULL);
-	uint32_t *schedule = (uint32_t *) *_schedule;
-	aes_keysched_no_bitslice(schedule, (const unsigned char *) key, 32);
-}
-
 void oqs_aes128_ecb_enc_sch_c(const uint8_t *plaintext, const size_t plaintext_len, const void *schedule, uint8_t *ciphertext) {
 	assert(plaintext_len % 16 == 0);
 	const aes128ctx *ctx = (const aes128ctx *) schedule;
@@ -749,6 +753,7 @@ void oqs_aes128_free_schedule_c(void *schedule) {
 		OQS_MEM_secure_free(ctx, sizeof(aes128ctx));
 	}
 }
+
 void oqs_aes256_free_schedule_c(void *schedule) {
 	if (schedule != NULL) {
 		aes256ctx *ctx = (aes256ctx *) schedule;
@@ -756,19 +761,14 @@ void oqs_aes256_free_schedule_c(void *schedule) {
 	}
 }
 
+void oqs_aes256_free_schedule_no_bitslice(void *schedule) {
+	if (schedule != NULL) {
+		OQS_MEM_secure_free(schedule, sizeof(aes256ctx_nobitslice));
+	}
+}
+
 void oqs_aes128_free_schedule_no_bitslice(void *schedule) {
 	if (schedule != NULL) {
 		OQS_MEM_secure_free(schedule, 44 * sizeof(int));
 	}
-}
-
-void oqs_aes256_free_schedule_no_bitslice(void *schedule) {
-	if (schedule != NULL) {
-		OQS_MEM_secure_free(schedule, 60 * sizeof(int));
-	}
-}
-
-void oqs_aes256_ctr_enc_sch_upd_blks_u64_c(uint64_t iv, void *ctx, uint8_t *out, size_t out_blks) {
-	oqs_aes256_load_iv_u64_c(iv, ctx);
-	oqs_aes256_ctr_enc_sch_upd_blks_c(ctx, out, out_blks);
 }
